Refactor query endpoints into new file
